feat: implement new workspace layout with Kanban board integration and improved navigation
