+++ src/app.html
@@ -5,7 +5,7 @@
   <meta charset="utf-8" />
   <link rel="icon" href="%sveltekit.assets%/favicon.png" />
   <meta name="viewport" content="width=device-width, initial-scale=1" />
-  <link rel="stylesheet" href="css/pico.min.css">
+  <link rel="stylesheet" href="/css/pico.min.css">
   <title>Tauri + SvelteKit + Typescript App</title>
   %sveltekit.head%
 </head>
