+++ src/routes/workspace/boards/[board_id]/+page.ts
@@ -1,0 +1,5 @@
+import type { PageLoad } from "./$types";
+
+export const load: PageLoad = ({ params }) => ({
+    board_id: params.board_id,
+});
